refactor(api): Move Telestion verticle with traits into own package
