+++ modules/telestion-api/src/main/java/de/wuespace/telestion/api/verticle/TelestionConfiguration.java
@@ -1,4 +1,4 @@
-package de.wuespace.telestion.api;
+package de.wuespace.telestion.api.verticle;
 
 import de.wuespace.telestion.api.message.JsonMessage;
