+++ modules/telestion-example/src/main/java/de/wuespace/telestion/example/PositionPublisher.java
@@ -1,7 +1,7 @@
 package de.wuespace.telestion.example;
 
-import de.wuespace.telestion.api.TelestionVerticle;
-import de.wuespace.telestion.api.traits.WithEventBus;
+import de.wuespace.telestion.api.verticle.TelestionVerticle;
+import de.wuespace.telestion.api.verticle.traits.WithEventBus;
 import de.wuespace.telestion.example.messages.Position;
 import io.vertx.core.Vertx;
